backend/requirements.txt
google-cloud-storage
Flask==1.1.1
Keras==2.3.1
numpy==1.18.1
opencv-python==4.2.0.32
Pillow==7.0.0
scipy==1.4.1
tensorflow-cpu==1.15.0
Werkzeug==1.0.0

